+++ src/org/mapyrus/ContextStack.java
@@ -181,7 +181,7 @@ public class ContextStack
 	 * @param stdoutStream standard output stream for program.
 	 */
 	public void setOutputFormat(String format, String filename,
-		int width, int height, String extras,
+		double width, double height, String extras,
 		PrintStream stdoutStream)
 		throws IOException, MapyrusException
 	{
